+++ templatevariantmappingitem.go
@@ -2,5 +2,5 @@ package model
 
 type TemplateVariantMappingItem struct {
 	Placement  string `json:"placement" bson:"placement"`
-	TemplateID int    `json:"template_id" bson:"template_id"`
+	TemplateID int64  `json:"template_id" bson:"template_id"`
 }
